InSighto/requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
matplotlib==3.8.2
requests==2.31.0
numpy==1.26.2
seaborn==0.13.0
flask-sqlalchemy
flask-login
werkzeug
gunicorn
python-dotenv